feat(settings): People section combining Users & Crew + Notifications (#228)
Ship Slice 2 of the Settings redesign (PRD #226):

- /settings/people: two-tab shell — Users & Crew, Notifications — built on
  the SettingsTabs primitive from #227. Tab bodies extracted into
  co-located *-tab.tsx components (content unchanged).
- Notifications tab subtitle now flags its per-user nature
  ("Personal setting — these preferences only affect you, not other
  people in your organization.") so admins don't assume it's org-wide.
  Header moved outside the loading branch so the disclosure shows
  while preferences fetch.
- Server-side redirects from /settings/{users,notifications} to
  /settings/people?tab=<key> so bookmarks and the notification-bell
  deep link keep working.
- settings-nav.ts: one new People entry (icon: Users); the two old
  entries are removed. Mobile dropdown picks this up automatically
  (same source of truth).

Tests:
- src/app/settings/people/page.test.tsx — shell smoke
- src/app/settings/people/notifications-tab.test.tsx — per-user disclosure
- src/lib/settings-nav.test.ts — entry swap
- src/app/settings/{users,notifications}/page.test.tsx — redirect targets
