+++ src/lib/settings-nav.ts
@@ -7,7 +7,6 @@ import {
   Users,
   Mail,
   FileSignature,
-  Bell,
   FileText,
   Download,
   Send,
@@ -42,7 +41,7 @@ export const settingsNavItems: SettingsNavItem[] = [
     label: "Expense Categories",
     icon: Receipt,
   },
-  { href: "/settings/users", label: "Users & Crew", icon: Users },
+  { href: "/settings/people", label: "People", icon: Users },
   { href: "/settings/email", label: "Email Accounts", icon: Mail },
   {
     href: "/settings/signatures",
@@ -55,7 +54,6 @@ export const settingsNavItems: SettingsNavItem[] = [
     icon: FileText,
   },
   { href: "/settings/contracts", label: "Contracts", icon: Send },
-  { href: "/settings/notifications", label: "Notifications", icon: Bell },
   { href: "/settings/accounting", label: "Accounting", icon: Link2 },
   { href: "/settings/stripe", label: "Stripe Payments", icon: CreditCard },
   { href: "/settings/payments", label: "Outgoing Emails", icon: Mail },
